se arreglo los modals y detalles pocos

+++ myapp/views.py
@@ -10,9 +10,6 @@ from .forms import (CreateNewCursos,
                     validar_delete_cursos,
                     validar_delete_public)
 from .models import create_nuevo_curso, create_nuevo_card , create_nuevo_publicacion
-from django.http import HttpRequest
-import base64
-from django.core.files.storage import FileSystemStorage
 from django.conf import settings
 # Create your views here.
 
@@ -20,12 +17,18 @@ from django.conf import settings
 def vistageneral(request):
     perfil_card = create_nuevo_card.objects.all()
     publicidad = create_nuevo_publicacion.objects.all()
-    return render(request, 'index/index.html',{"perfil_card": perfil_card , "publicidad": publicidad})
+    return render(request, 'index/index.html',{"perfil_card": perfil_card, "publicidad": publicidad})
 
 
 def lista_cursos_funcionarios(request):
     cursos = create_nuevo_curso.objects.all()
-    return render(request, 'index/lista_cursos_fun.html',{"cursos":cursos})
+    for lista in cursos:
+        if lista != None:
+            return render(request, 'index/lista_cursos_fun.html',{"cursos":cursos })
+    return render(request, 'index/lista_cursos_fun.html')
+
+def form_consulta(request):
+    return render(request, 'index/consulta_cursos.html')
 
 
 def login_operador(request):
@@ -41,25 +44,13 @@ def modificar_index(request):
 def ingresar_cursos(request):
     return render(request, 'operador/cursos_add.html')
 
-
 def lista_cursos_operador(request):
     cursos = create_nuevo_curso.objects.all()
     for lista in cursos:
         if lista != None:
-            arr = lista.personal_cursos.split(" ")
-            lista_2 = []
-            for x in arr:
-                lista_2.append(x)
-        return render(request, 'operador/lista_cursos.html',{"cursos":cursos, "lista_2":lista_2})
+            return render(request, 'operador/lista_cursos.html',{"cursos":cursos})
     return render(request, 'operador/lista_cursos.html')
 
-
-
-#Actualizar tablas
-def actualizar_tabla_modindex():
-    perfil_card = create_nuevo_card.objects.all()
-    publicidad = create_nuevo_publicacion.objects.all()
-    return {"perfil": perfil_card , "publicidad": publicidad}
 
 #list a str
 def convertidor_list_str(list):
@@ -100,13 +91,15 @@ def form_create_new_card(request):
     if request.method == 'POST':
         form = CreateNewCard(request.POST, request.FILES)
         if form.is_valid():
+            perfil_card = create_nuevo_card.objects.all()
+            publicidad = create_nuevo_publicacion.objects.all()
             model_card = create_nuevo_card()
             model_card.img_card = request.FILES['img_card']
             model_card.nombre_card = request.POST['nombre_card']
             model_card.cargo_card = request.POST['cargo_card']
             model_card.descrip_card = request.POST['descrip_card']
             model_card.save()
-            return render(request, "operador/mod_index.html")
+            return render(request, 'operador/mod_index.html',{"perfil_card": perfil_card , "publicidad": publicidad})
         else: 
             return render(request, "operador/mod_index.html", {"ventana": True})
     else:
@@ -118,12 +111,14 @@ def form_create_new_publicacion(request):
     if request.method == 'POST':
         form = CreateNewPublication(request.POST, request.FILES)
         if form.is_valid():
+            perfil_card = create_nuevo_card.objects.all()
+            publicidad = create_nuevo_publicacion.objects.all()
             model_pub = create_nuevo_publicacion()
             model_pub.img_pub = request.FILES['img_pub']
             model_pub.nombre_pub = request.POST['nombre_pub']
             model_pub.descrip_pub = request.POST['descrip_pub']
             model_pub.save()
-            return render(request, "operador/mod_index.html")
+            return render(request, 'operador/mod_index.html',{"perfil_card": perfil_card , "publicidad": publicidad,'ventana_public':True})
         else:
             return render(request, "operador/mod_index.html", {"ventana": True})
     else:
@@ -137,8 +132,10 @@ def delete_objet_curso(request):
         if form.is_valid():
             id_curso = request.POST['id_delete_curso']
             instance = create_nuevo_curso.objects.get(id=id_curso)
-            os.remove("myapp"+instance.img_curso.url)
-            os.remove("myapp"+instance.form_carga_curso.url)
+            if os.path.isfile("myapp"+instance.img_curso.url) == True:
+                os.remove("myapp"+instance.img_curso.url)
+            if  os.path.isfile("myapp"+instance.form_carga_curso.url) == True:
+                os.remove("myapp"+instance.form_carga_curso.url)
             instance.delete()
             return redirect("/lista_cursos_operador/")
         
@@ -146,19 +143,28 @@ def delte_objet_card(request):
     if request.method == 'POST':
         form = validar_delete_card(request.POST)
         if form.is_valid():
+            perfil_card = create_nuevo_card.objects.all()
+            publicidad = create_nuevo_publicacion.objects.all()
             id_card = request.POST['id_delete_card']
             instance = create_nuevo_card.objects.get(id=id_card)
-            os.remove("myapp"+instance.img_card.url)
+            if os.path.isfile("myapp"+instance.img_card.url) == True:
+                os.remove("myapp"+instance.img_card.url)
             instance.delete()
-            return redirect("/mod_index/")
+            return render(request, 'operador/mod_index.html',{"perfil_card": perfil_card , "publicidad": publicidad})
     
 def delte_objet_public(request):
+    perfil_card = create_nuevo_card.objects.all()
+    publicidad = create_nuevo_publicacion.objects.all()
     if request.method == 'POST':
         form = validar_delete_public(request.POST)
         if form.is_valid():
             id_public = request.POST['id_delete_public']
             instance = create_nuevo_publicacion.objects.get(id=id_public)
-            os.remove("myapp"+instance.img_pub.url)
+            if os.path.isfile("myapp"+instance.img_pub.url) == True:
+                os.remove("myapp"+instance.img_pub.url)
             instance.delete()
-            return redirect("/mod_index/")
-    
+            return render(request, 'operador/mod_index.html',{"perfil_card": perfil_card , "publicidad": publicidad,'ventana_public_delete':True})
+        else:
+            return render(request, 'operador/mod_index.html',{"perfil_card": perfil_card , "publicidad": publicidad,'ventana_public_delete_f':True})
+    else:
+        return render(request, 'operador/mod_index.html',{"perfil_card": perfil_card , "publicidad": publicidad,'ventana_public_delete_x':True})
